+++ packages/editor/src/store/test/reducer.js
@@ -18,6 +18,8 @@ import {
 	saving,
 	reusableBlocks,
 	autosave,
+	template,
+	blockListSettings,
 	postSavingLock,
 	previewLink,
 } from '../reducer';
@@ -798,38 +800,6 @@ describe( 'state', () => {
 				data: {},
 				isFetching: {},
 				isSaving: {},
-			} );
-		} );
-	} );
-
-	describe( 'autosave', () => {
-		it( 'returns null by default', () => {
-			const state = autosave( undefined, {} );
-
-			expect( state ).toBe( null );
-		} );
-
-		it( 'returns subset of received autosave post properties', () => {
-			const state = autosave( undefined, {
-				type: 'RESET_AUTOSAVE',
-				post: {
-					title: {
-						raw: 'The Title',
-					},
-					content: {
-						raw: 'The Content',
-					},
-					excerpt: {
-						raw: 'The Excerpt',
-					},
-					status: 'draft',
-				},
-			} );
-
-			expect( state ).toEqual( {
-				title: 'The Title',
-				content: 'The Content',
-				excerpt: 'The Excerpt',
 			} );
 		} );
 	} );